removed maxint

+++ src/Coord.java
@@ -9,6 +9,6 @@ public class Coord {
 
   @Override
   public String toString() {
-    return "("+x+","+y+")";
+    return "("+(int)x+","+(int)y+")";
   }
 }
